Enable local timezone
Unless you specify a timezone in synology, the docker container will always use UTC.
Since the app uses `localtime` conversion, a timezone has to be set in order to work properly.
This is of course personal preference (Europe/Berlin), but from what I got who is using this image, I expect it to fit most of the times.
Adapt to your needs.